# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Clinic owner authentication
  - name: dashboard
    description: Clinic operations overview
  - name: inbox
    description: Omnichannel inbox and messaging
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/session:
    get:
      operationId: getAuthSession
      tags: [auth]
      summary: Get the current clinic user session
      responses:
        "200":
          description: Current authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in a clinic owner or admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: recoverPassword
      tags: [auth]
      summary: Send a password recovery email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordRecoveryInput"
      responses:
        "200":
          description: Recovery request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordRecoveryResponse"
        "400":
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Set a new password using a recovery token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetResponse"
        "400":
          description: Invalid or expired recovery token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new clinic owner account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered clinic owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Registration failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Sign out the current user
      responses:
        "204":
          description: Signed out
  /settings:
    get:
      operationId: getClinicSettings
      tags: [auth]
      summary: Get the current clinic settings
      responses:
        "200":
          description: Current clinic settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClinicSettings
      tags: [auth]
      summary: Update the current clinic settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClinicSettingsInput"
      responses:
        "200":
          description: Updated clinic settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Invalid settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get the clinic dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inbox/inbound:
    post:
      operationId: receiveInboundMessage
      tags: [inbox]
      summary: Webhook receiver for incoming patient messages from external channels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InboundMessageInput"
      responses:
        "201":
          description: Message processed and stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboundMessageResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
      required: [email, password]
    PasswordRecoveryInput:
      type: object
      properties:
        email:
          type: string
      required: [email]
    PasswordRecoveryResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ResetPasswordInput:
      type: object
      properties:
        accessToken:
          type: string
          minLength: 20
        password:
          type: string
          minLength: 8
      required: [accessToken, password]
    PasswordResetResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    RegisterInput:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
        clinicName:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 8
      required: [fullName, clinicName, email, password]
    AuthSession:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/ClinicUser"
        clinic:
          $ref: "#/components/schemas/Clinic"
      required: [user, clinic]
    ClinicUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [owner, admin]
      required: [id, email, fullName, role]
    Clinic:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
        city:
          type: string
      required: [id, name, status, city]
    UpdateClinicSettingsInput:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 160
        clinicName:
          type: string
          minLength: 2
          maxLength: 160
        city:
          type: string
          maxLength: 120
      required: [fullName, clinicName, city]
    DashboardSummary:
      type: object
      properties:
        clinic:
          $ref: "#/components/schemas/Clinic"
        stats:
          type: array
          items:
            $ref: "#/components/schemas/DashboardStat"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required: [clinic, stats, recentActivity]
    DashboardStat:
      type: object
      properties:
        label:
          type: string
        value:
          type: string
        helper:
          type: string
        tone:
          type: string
          enum: [blue, green, amber, violet]
      required: [label, value, helper, tone]
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        createdAt:
          type: string
        tone:
          type: string
          enum: [blue, green, amber, violet]
      required: [id, title, description, createdAt, tone]
    InboundMessageInput:
      type: object
      properties:
        clinicId:
          type: string
        channelType:
          type: string
          enum: [whatsapp, instagram, messenger, telegram, sms, web]
        channelId:
          type: string
        senderPhone:
          type: string
        senderName:
          type: string
        content:
          type: string
          minLength: 1
      required: [content]
    InboundMessageResponse:
      type: object
      properties:
        ok:
          type: boolean
        messageId:
          type: string
        conversationId:
          type: string
        patientId:
          type: string
        isHandoff:
          type: boolean
      required: [ok, messageId, conversationId]