# lib/api-zod/src/generated/types/inboundMessageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InboundMessageInputChannelType } from './inboundMessageInputChannelType';

export interface InboundMessageInput {
  clinicId?: string;
  channelType?: InboundMessageInputChannelType;
  channelId?: string;
  senderPhone?: string;
  senderName?: string;
  /** @minLength 1 */
  content: string;
}