# lib/api-zod/src/generated/types/inboundMessageResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InboundMessageResponse {
  ok: boolean;
  messageId: string;
  conversationId: string;
  patientId?: string;
  isHandoff?: boolean;
}